Create maverrors.py to resolve circular imports
In pymavlink/generator/mavparse.py from .mavgen_python import MAVError to resolve undefined name which can raise `NameError` at runtime.

Resolve the circular imports between mavgen_python.py, mavparse.py, and mavtemplate.py by putting MAVError and MAVParseError into their own maverrors.py file.

Fixes #116,#117
